+++ platformio/__init__.py
@@ -34,3 +34,4 @@ __license__ = "Apache Software License"
 __copyright__ = "Copyright 2014-present PlatformIO"
 
 __apiurl__ = "https://api.platformio.org"
+__pioremote_endpoint__ = "ssl:remote.platformio.org:4413"